[Development] Loading the keyframe data, using HTTP Range headers to load only the desired part of the .fbk file. In this revision, the data then cannot be parsed correctly because RfbProto expects RFB initialization sequence which is absent in keyframes.
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@2609 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ java/src/com/tightvnc/rfbplayer/FbsInputStream.java
@@ -43,7 +43,7 @@ class FbsInputStream extends InputStream {
   protected int bufferPos;
 
   /** The number of bytes to skip in the beginning of the next data block. */
-  protected int nextBlockOffset;
+  protected long nextBlockOffset;
 
   protected Observer obs;
 
@@ -95,7 +95,7 @@ class FbsInputStream extends InputStream {
    *    from <code>in</code>.
    */
   FbsInputStream(InputStream in, long timeOffset, byte[] buffer,
-                 int nextBlockOffset) {
+                 long nextBlockOffset) {
 
     this.in = in;
     startTime = System.currentTimeMillis() - timeOffset;
